[Xamarin.Android.Build.Tasks] fix <CilStrip/> for Hybrid AOT (#4850)
Fixes: https://github.com/xamarin/xamarin-android/issues/4818

Use of Hybrid AOT currently results in assemblies that have not been
CIL-stripped.

In 77ab2404, a change was made to fix how `Mono.Android.dll` was
treated during AOT.  `Mono.Android.dll` used to be the only assembly
copied to the `shrunk` directory, and we had two copies of the this
assembly passed to the AOT compiler.  This change unintentionally
made `<CilStrip/>` strip the wrong set of assemblies...

The initial fix would be:

	<CilStrip
	    ...
	-   ResolvedAssemblies="@(_ResolvedAssemblies)">
	+   ResolvedAssemblies="@(_ShrunkAssemblies)">

Unfortunately, this causes a crash:

	E/AndroidRuntime(18806): java.lang.UnsatisfiedLinkError: No implementation found for void mono.android.TypeManager.n_activate(java.lang.String, java.lang.String, java.lang.Object, java.lang.Object[]) (tried Java_mono_android_TypeManager_n_1activate and Java_mono_android_TypeManager_n_1activate__Ljava_lang_String_2Ljava_lang_String_2Ljava_lang_Object_2_3Ljava_lang_Object_2)
	E/AndroidRuntime(18806): 	at mono.android.TypeManager.n_activate(Native Method)
	E/AndroidRuntime(18806): 	at mono.android.TypeManager.Activate(:7)
	E/AndroidRuntime(18806): 	at crc64446c24ccf511bf5f.SplashScreenActivity.<init>(:25)
	E/AndroidRuntime(18806): 	at java.lang.Class.newInstance(Native Method)
	E/AndroidRuntime(18806): 	at android.app.Instrumentation.newActivity(Instrumentation.java:1086)
	E/AndroidRuntime(18806): 	at android.app.ActivityThread.performLaunchActivity(ActivityThread.java:2809)
	E/AndroidRuntime(18806): 	at android.app.ActivityThread.handleLaunchActivity(ActivityThread.java:2988)
	E/AndroidRuntime(18806): 	at android.app.ActivityThread.-wrap14(ActivityThread.java)
	E/AndroidRuntime(18806): 	at android.app.ActivityThread$H.handleMessage(ActivityThread.java:1631)
	E/AndroidRuntime(18806): 	at android.os.Handler.dispatchMessage(Handler.java:102)
	E/AndroidRuntime(18806): 	at android.os.Looper.loop(Looper.java:154)
	E/AndroidRuntime(18806): 	at android.app.ActivityThread.main(ActivityThread.java:6682)
	E/AndroidRuntime(18806): 	at java.lang.reflect.Method.invoke(Native Method)
	E/AndroidRuntime(18806): 	at com.android.internal.os.ZygoteInit$MethodAndArgsCaller.run(ZygoteInit.java:1520)
	E/AndroidRuntime(18806): 	at com.android.internal.os.ZygoteInit.main(ZygoteInit.java:1410)

Thinking about how things "used to work", we weren't stripping the
*correct* `Mono.Android.dll`.  So we could run `<CilStrip/>` on every
assembly *besides* `Mono.Android.dll`?

I think this could be improved further if we could strip
`Mono.Android.dll`, but this at least gets things back to working the
way they used to.

I updated the `BuildTest.BuildIncrementalAot()` test that had a
`//TODO` comment, as it works properly now.  I also added a new test
to check that method bodies are stripped.