[Xamarin.Android.Build.Tasks] don't delete ILMerge'd assemblies (#4836)
Context: https://build.azdo.io/3823945

The monodroid build is failing with:

	create-pkg.targets(34,5): error MSB3030: Could not copy the file "…/external/xamarin-android/bin/Debug/lib/xamarin.android/xbuild/Xamarin/Android/System.Collections.Immutable.dll" because it was not found.
	    0 Warning(s)
	    1 Error(s)

We added this file to `create-installers.targets` in 2f677efa, because
it is needed for the new "delta install" support.

However, there is an issue depending on the ordering of how things are
built.  If `Xamarin.Android.Build.Tasks.csproj` is built *last*, it
will `ILMerge` this assembly and delete it.

Let's not delete the ILMerge'd assemblies, which avoids the MSB3030.