[runtime] Preload `libmono-native.so` before initializing XA
Fixes: https://github.com/xamarin/xamarin-android/issues/4772
Fixes: https://github.com/xamarin/xamarin-android/issues/4852

Fixes apply to the the **Release** builds on Android 5.x devices.

Android 5.x devices have a bug in their 64-bit dynamic linker which
prevents it from correctly resolving shared library dependencies when a
library required by the one currently being loaded is not found in the
same directory.  The problem is that the Android dynamic linker on those
devices fails to properly interpret the internal library name (so-called
`SONAME`, contained in the `DT_SONAME` field in the ELF executable
header) when checking whether the library is already loaded in the
process memory.  Failing that check, the loader proceeds to look for the
library **file** on the disk/storage medium in the predefined paths.
Since `libmonosgen-2.0.so` and other libraries are not found in those
paths, the search fails and the linker returns an error.  Later versions
of Android fix the bug in that the `DT_SONAME` field is used to properly
look up already loaded shared libraries and thus the dependency
resolution succeeds.

The solution here is to use the Java API, `System.loadLibrary`, to
preload `libmono-native.so` because the Java API is able to somehow
figure out that the dependent libraries are found in the same directory
and load them to satisfy the requirements.  We could probably implement
the same algorithms in our native code, but it feels it would be a
wasted effort since Android versions above 5.x work correctly.