[generator] Mark protected types nested in sealed classes as private (#672)
In commit 07555606 we marked `protected` members in `sealed` classes
as `private` in order to prevent `csc` warnings.  However, we did
address nested types which can also cause this warning:

	sealed class Foo
	{
	    protected interface Bar
	    {
	    }

	    protected class Bar2
	    {
	    }
	}

which results in CS0628 warnings:

	warning CS0628: 'Foo.Bar': new protected member declared in sealed class
	warning CS0628: 'Foo.Bar2': new protected member declared in sealed class

`Mono.Android.dll` in API-30 added 28 new protected interfaces in the
`sealed` classes `CalendarContract` and `ContactContract`, which
introduced new warnings in our builds:


	obj\Release\monoandroid10\android-30\mcw\Android.Provider.CalendarContract.cs(614,40): warning CS0628: 'CalendarContract.IAttendeesColumns': new protected member declared in sealed class

Avoid the CS0628 warnings by marking nested types as `private` if the
parent class is `sealed`.