[tests] Disable a Mono.Posix test
`Android 11` / `API 30` made the `memfd_create(2)` Linux-specific API
fail whenever the `MFD_CLOEXEC | MFD_ALLOW_SEALING` combination of flags
is passed to it:

    Xamarin.Android.Bcl_Tests, MonoTests.Mono.Unix.Native.MemfdTest.TestMemfd / Release

    System.ArgumentOutOfRangeException : Current platform doesn't support this value.
    Parameter name: value
    Actual value was MFD_CLOEXEC, MFD_ALLOW_SEALING.

Disable the test altogether, since this API isn't used by
`Xamarin.Android` and any applications using it will/should check the
`errno` variable for the `ENOSYS` value to see if the API is supported
or not.  This will unblock us, allowing update of our Android emulator
system image to API 30.