Bump $(ProductVersion) to 11.0.99
Context: https://developer.android.com/android11

We (try to) base Xamarin.Android product versions on the latest
Android versions, and API-30 is "Android 11".  With the merging of
stable API-30 support (2181e786), we want our first release with
API-30 support to likewise be v11.0.

The [d16-7][0] branch is the first release which will offer stable
API-30 support, and was bumped to report itself as [v11.0][1].

As xamarin-android/master is tracking the *next* Xamarin.Android
version, bump `$(ProductVersion)` to v11.0.99 for eventual release
as v11.1.0, and bump `$(AndroidPackVersion)` to `v11.0.100`.

[0]: https://github.com/xamarin/xamarin-android/commits/d16-7
[1]: https://github.com/xamarin/xamarin-android/commit/dc830f7e81ca5c17c438cf39c5953df9f99dd452