[Xamarin.Android.Build.Tasks] Set LatestSupportedJavaVersion=v11.0
Fixes: https://github.com/xamarin/xamarin-android/issues/4853

`$(LatestSupportedJavaVersion)` is currently v11.0.4, which means if
e.g. JDK 11.0.7 is used -- which one would expect to be compatible! --
the build will fail with an XA0030 error:

	error XA0030: Building with JDK Version `11.0.7` is not supported. Please install JDK version `11.0.4`.

We *could* set `$(LatestSupportedJavaVersion)`=v11.0.99, but that
value is also used as part of the XA0030 message text, which could
result in an error message mentioning a version which doesn't exist:

	error XA0030: Building with JDK Version `14.0.0` is not supported. Please install JDK version `11.0.99`.

Instead of using a three-tuple in the `$(LatestSupportedJavaVersion)`
value -- which is "forwarded" to the XA0030 error -- instead use a
two-tuple value of `11.0`, which will result in a "saner" XA0030 error
message of `11.0` and not `11.0.4` or (worse!) `11.0.99`.

To bridge the gap between three-tuples (from `java -version`) and
the two-tuple in `$(LatestSupportedJavaVersion)`, add a new
`ValidateJavaVersion.CheckJavaGreaterThanLatestSupported()` method
which will *append* `.99` to `$(LatestSupportedJavaVersion)` if it
only contains two values, allowing us to compare `11.0.7 >= 11.0.99`.