[Xamarin.Android.Build.Tasks] support JDK 11.0.7
Fixes: https://github.com/xamarin/xamarin-android/issues/4853
Context: https://confluence.jetbrains.com/display/JBR/JetBrains+Runtime

Using JDK 11.0.7, builds currently fail with:

    error XA0030: Building with JDK version `11.0.7` is not supported. Please install JDK version `11.0.4`. See https://aka.ms/xamarin/jdk9-errors

We could set `$(LatestSupportedJavaVersion)` to `11.0.7` to allow this
version to work, but it is not future-proof.

So if we used `11.0.99`, this would allow Xamarin.Android to continue
to work with minor JDK bumps if 11.0.8 happened to come out tomorrow.

Unfortunately, this makes the error message bad:

    error XA0030: Building with JDK version `14.0.0` is not supported. Please install JDK version `11.0.99`.

I updated the `<ValidateJavaVersion/>` MSBuild task, so that only the
first two components of the version appear in this message:

    error XA0030: Building with JDK version `14.0.0` is not supported. Please install JDK version `11.0`.