[One .NET] fix debugging (#4864)
If you tried to debug an app with:

	dotnet build HelloAndroid/HelloAndroid.csproj -t:Install,_Run -p:AndroidAttachDebugger=true

The presence of `$(AndroidAttachDebugger)` causes a crash on startup:

	E mono    : Unhandled Exception:
	E mono    : System.ArgumentNullException: Value cannot be null. (Parameter 'meth')
	E mono    :    at System.Reflection.Emit.ILGenerator.Emit(OpCode opcode, MethodInfo meth)
	E mono    :    at Android.Runtime.JNINativeWrapper.CreateDelegate(Delegate dlg)
	E mono    :    at Android.App.Activity.GetOnCreate_Landroid_os_Bundle_Handler()
	E mono    :    at Android.Runtime.AndroidTypeManager.RegisterNativeMembers(JniType nativeClass, Type type, String methods)
	E mono    :    at Android.Runtime.JNIEnv.RegisterJniNatives(IntPtr typeName_ptr, Int32 typeName_len, IntPtr jniClass, IntPtr methods_ptr, Int32 methods_len)
	E mono-rt : [ERROR] FATAL UNHANDLED EXCEPTION: System.ArgumentNullException: Value cannot be null. (Parameter 'meth')
	E mono-rt :    at System.Reflection.Emit.ILGenerator.Emit(OpCode opcode, MethodInfo meth)
	E mono-rt :    at Android.Runtime.JNINativeWrapper.CreateDelegate(Delegate dlg)
	E mono-rt :    at Android.App.Activity.GetOnCreate_Landroid_os_Bundle_Handler()
	E mono-rt :    at Android.Runtime.AndroidTypeManager.RegisterNativeMembers(JniType nativeClass, Type type, String methods)
	E mono-rt :    at Android.Runtime.JNIEnv.RegisterJniNatives(IntPtr typeName_ptr, Int32 typeName_len, IntPtr jniClass, IntPtr methods_ptr, Int32 methods_len)

The cause of the crash is because we pass
`JNINativeWrapper.mono_unhandled_exception_method` to
[`ILGenerator.Emit(OpCode, MethodInfo)`][0], and
`mono_unhandled_exception_method` is `null`.

`mono_unhandled_exception_method` is a `MethodInfo` to
[`Debugger.Mono_UnhandledException()`][1], which does not exist in
.NET 5 Mono, which is why it's `null` within .NET 5.

To fix `dotnet build -t:_Run` support, cleanup
`JNINativeWrapper.get_runtime_types()` to instead check for
`JNINativeWrapper.exception_handler_method`, as
`exception_handler_method` is consistently initialized properly on
both "legacy" and .NET 5 environments.

Furthermore, update `get_runtime_types()` so that
`mono_unhandled_exception_method` is not passed to `ILGenerator.Emit()`
when it is `null`.  This avoids the `ArgumentNullException`.

This surfaced another problem when using the `_Run` target by itself:

	Xamarin.Android.Common.Debugging.targets(518,2): error MSB3073: The command ""adb"  forward tcp:10000 tcp:10000" exited with code 9009.

The `$(_AndroidPlatformToolsDirectory)` property was blank when running
the `_Run` target by itself, but things work fine if you run
`Install,_Run` at the same time.

This behavior broke in commit fb637e08.  I accidentally replaced the
`AndroidPrepareForBuild` MSBuild target with [the version][2] in
`Xamarin.Android.Bindings.Core.targets`.

To fix this, I can just move the binding project version of
`AndroidPrepareForBuild` target to `Xamarin.Android.Bindings.targets`.

I added a new test verifying that breakpoints work, when running
under .NET 5.

[0]: https://docs.microsoft.com/en-us/dotnet/api/system.reflection.emit.ilgenerator.emit?view=netcore-3.1#System_Reflection_Emit_ILGenerator_Emit_System_Reflection_Emit_OpCode_System_Reflection_MethodInfo_
[1]: https://github.com/mono/mono/blob/2ff424be2933f711207a8139c5792ab5d1b495a9/mcs/class/corlib/System.Diagnostics/Debugger.cs#L119-L127
[2]: https://github.com/xamarin/xamarin-android/blob/f99a3fad1b12d9674de2f624e243f296cfc40fc6/src/Xamarin.Android.Build.Tasks/MSBuild/Xamarin/Android/Xamarin.Android.Bindings.Core.targets#L47