[Microsoft.Android.Sdk] fixes for ILLink and @(ProguardConfiguration)
When using r8 or ProGuard in .NET 5, `proguard_project_references.cfg`
or `$(_ProguardProjectConfiguration)` was not being generated. It
turns out that the `$(_ProguardProjectConfiguration)` MSBuild property
was completely blank.

`$(_ProguardProjectConfiguration)` is used when deciding if
r8/ProGuard need to run such as:

https://github.com/xamarin/xamarin-android/blob/a944a8923c0f5e782af308525a3e99ebc2470fe8/src/Xamarin.Android.Build.Tasks/Xamarin.Android.D8.targets#L28

The way the inner/outer builds work in .NET 5+, we need to set
`$(_ProguardProjectConfiguration)` from the outer build and pass it
through to the inner build running the `ILLink` MSBuild target.

After this fix, I could enable the `BuildProguardEnabledProject` test
under .NET 5. The only thing to note is that the `*.widget.Toolbar`
class is the AndroidX version due to a newer Xamarin.Forms used:

https://github.com/xamarin/xamarin-android/blob/a944a8923c0f5e782af308525a3e99ebc2470fe8/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Android/XamarinFormsAndroidApplicationProject.cs#L43-L48

Xamarin.Forms 4.5.x will bring in AndroidX if `$(TFV)` >= `v10.0`.