[Microsoft.Android.Sdk] set $(PublishTrimmed) after .csproj evaluated (#4888)
I noticed that if you define `<AndroidLinkMode>Full</AndroidLinkMode>`
in a `.csproj` file, the `ILLink` MSBuild target did not run.

It turns out that we were setting `$(PublishTrimmed)` by default too
early. If set in `Microsoft.Android.Sdk.props`, this is evaluated
before the user's `.csproj` file. If we set this in
`Microsoft.Android.Sdk.targets`, then we can rely on values set in the
`.csproj` file.

I found a test that was failing due to this issue and added it to the
`dotnet` category. I made a couple other changes to the test so it
passed under .NET 5+:

* Test against `FormsViewGroup.dll` instead of a support library, so
  the test still passes if using AndroidX libraries.
* Use `$(RuntimeIdentifiers)` over `$(AndroidSupportedAbis)`. I added
  a new helper method for multiple RIDs.