[Xamarin.Android.Build.Tools] Import the new Xamarin.Android.Tools.Versions.props file.
In order to ensure that both xamarin.android and the sdk installer
use the same version number we are now using the `Xamarin.Android.Tools.Versions.props`
file. This file is maintained in the xamarin-android-tools repo and
will be imported via `external/xamarin-android-tools`. The file
is imported in `Xamarin.Android.Common.props`.