Work around an Android 9 dlopen bug on x86
Fixes: https://github.com/xamarin/xamarin-android/issues/4893

This commit fixes an issue in Android 9 (API 28) 32-bit (ONLY) dynamic
loader issue where if the same library is loaded more than once, instead
of returning a handle to the already loaded copy, `dlopen` will
reinitialize the library thus resetting its state (BSS section is
cleared, global instances of C++ classes are re-instantiated etc).

When the library is reinitalized, however, our runtime is not - since
the Java `Runtime.initialize`, which ends up in our native code,
is *not* called because the library load is not part of the process
startup, when `Runtime.initialize` is invoked.  This in turn leads to
situation when certain class members initialized in `Runtime.initialize`
and assumed to be correct throughout the application life are, in fact,
invalid and lead to crashes.  In the instance of issue #4893, the
problem was that the `MonodroidRuntime::Class_getName` field was not
initialized to point to the `Java.Lang.Class.getName` Java method and
that, in turn, led to Xamarin.Android not being able to look up Java
types during application execution time.

`libmonodroid.so` is reloaded whenever Mono tries to resolve an
`__Internal` p/invoke (via a call to Xamarin.Android's
`monodroid_dlopen` fallback handler) - these p/invokes are assumed to
live in `libmonodroid` so we need to return a handle to it, so that Mono
can subsequently use `dlsym` to look up the required symbol.  Given the
situation described in the previous paragraph, we need a way to not
reload `libmonodroid` or do something else to avoid it being
reinitialized.

Since there's no portable way to check whether a DSO is already
loaded we need to resort to "something else", implemented in this
commit.  We move all the functions used in `__Internal` p/invokes to a
separate DSO which can be reloaded & reinitialized over and over again
without shredding our global state.  This is done by way of introducing
a C++ interface class, `MonoAndroidExternalAPI` which is passed to the
new DSO on load time, thus giving it a way to invoke all the necessary
internal Xamarin.Android functions the `__Internal` p/invokes need in
order to work as expected.