[build] Fix compression of assemblies in unwritable directories
Context: d236af54538ef1fe62d0125798cdde78e46dcd8e

Commit d236af54 introduced assembly compression for `Release` builds
with compression output placed in the same directory where the original
assembly resides.  This works well if the linker is enabled for the
project, because all the assemblies are copied to the `obj/` directory,
but if the linker is disabled, some assemblies will be read from their
original path, possibly a system location which is NOT writable by
non-root/non-admin processes.

This commit moves the compressed output location to a subdirectory of
the `obj/` directory for all the input assemblies.