[One .NET] fix the GetAndroidDependencies MSBuild target (#4934)
In a .NET 5+ project, the `GetAndroidDependencies` MSBuild target
would use `$(TargetFrameworkVersion)` (which is 5.0) and always report
that API 21 needs to be installed.

The `<CalculateProjectDependencies/>` MSBuild task should use
`$(_AndroidApiLevel)` if set, and fallback to using
`$(TargetFrameworkVersion)`.

Under .NET 5+, `$(_AndroidApiLevel)` will be set via different targets
running in `$(GetAndroidDependenciesDependsOn)`. `$(_AndroidApiLevel)`
will be blank in legacy projects.

I updated a test so it now passes under a `dotnet` context.