[build] PackDotNet target to simplify .NET 6 development (#4931)
When working on .NET 6, I use the following powershell function for
my "developer loop":

	function pack-xa-nugets([string] $configuration = 'Debug')
	{
	    & msbuild $xa\build-tools\xa-prep-tasks\xa-prep-tasks.csproj
	    & msbuild $xa\Xamarin.Android.sln -p:DisableApiCompatibilityCheck=true
	    & msbuild /t:CreateAllPacks build-tools/create-packs/Microsoft.Android.Sdk.proj
	    Remove-Item -r $xa\packages\microsoft.android.* -ErrorAction SilentlyContinue
	}

The steps here are:

 1. Build `xa-prep-tasks.csproj` so the `_CreateVersion` MSBuild target
    can run and [update git versioning information][0].

 2. Build `Xamarin.Android.sln`.

 3. Run `CreateAllPacks`, to create all the `.nupkg` files.

 4. Delete any previously extracted files in `packages/microsoft.android.*`

Step (4) is particularly important for sanity.  Otherwise, I have
attempted to build a project, but NuGet did not extract the new
`.nupkg` files because their versions were the same!

Because no one else has this script, I ported it to MSBuild so it can
live in this repo and work on all platforms.

On Windows:

	msbuild Xamarin.Android.sln /t:PackDotNet

On macOS:

	make pack-dotnet

Down the road we could improve this workflow further by copying files
directly.  The `PackDotNet` target currently takes about 40 seconds
to execute on my Windows desktop machine.

[0]: https://github.com/xamarin/xamarin-android/blob/9c880983914e2dbc3f384dea16b5351696732682/build-tools/xa-prep-tasks/xa-prep-tasks.targets#L4