[xaprepare] Redirect some git messages from stderr to stdout
`git` outputs its progress messages to the standard error stream which
`xaprepare` intercepts and prints on screen in red and in the following
format:

    stderr | Cloning into 'monodroid'...
    stderr | remote: Enumerating objects: 58, done
    stderr | Receiving objects:   0% (1/79000)
    stderr | Receiving objects:   1% (790/79000)
    stderr | Submodule 'external/Java.Interop' (https://github.com/xamarin/java.interop.git) registered for path 'external/Java.Interop'
    stderr | Submodule 'external/android-api-docs' (https://github.com/xamarin/android-api-docs) registered for path 'external/android-api-docs'

This may be confused for an actual error even though it is most
certainly a harmless progress output.  This commit adds code to
intercept the know progress messages and redirect them to standard
output using normal text formatting.  Should a git command exit with
error code, the `stderr` output is printed to the screen again, but this
time in red.