[monodroid] Use the "none" option for C++'s standard library (#4958)
Fixes: https://github.com/xamarin/xamarin-android/issues/4949
Context: https://github.com/xamarin/xamarin-android/pull/4952
Context: https://github.com/android/ndk/issues/744

Xamarin.Android uses the "system" version of the C++ standard
library (which is, in reality, a very basic stub containing only weak
definitions of the `new` and `delete` operator functions) in order to
avoid possible conflicts with Xamarin.Android applications which could
make use of their own version of the C++ standard library.  The
conflict would arise from the fact that the `libc++` library shipped
with the current versions of Android NDK is built in a way that would
expose a large number of its symbols which would conflict with any
other `libc++` used by the application.

Additionally, linking `libc++` into Xamarin.Android runtime would
result in a binary 4-6 times bigger than what we currently ship (when
linking statically) or by 6.2MB if the shared library version of
`libc++` were used.

The "system" version of `libstdc++` is deprecated in favor of `libc++`
and may be removed from the NDK in a near future.  In order to prepare
for this we need to switch to one of the supported STL implementations.
The static and dynamic linking options are not viable, as described
above, which leaves us with the `none` STL option.  However, using this
variety causes Android NDK's `cmake` toolchain to pass the
`-nostdlib++ -nostdinc++` options to the compiler.  The first of them
doesn't affect us, since we don't use any types from the `libc++`
library, but the second of them causes the standard C++ files, which we
do use, to not be found.

Use the `none` STL variety and work around the `-nostdinc++` flag
by detecting and specifying the include path to the standard C++
headers manually on the command line.  This allows us to no longer link
against `libstdc++.so`, while continuing to use "header-file only"
functionality.