[One .NET] Fix <InstallAndroidDependencies/> target and test
Context: https://github.com/xamarin/android-sdk-installer/pull/450

A fix for the `<InstallAndroidDependencies/>` target has been brought in
via the `android-sdk-installer` bump in xamarin/monodroid@b12a3e8d.

The `<InstallAndroidDependencies/>` test previously had issues due to
clean up logic in the test framework that partially deleted the newly
provisioned Android SDK on a subsequent build.  I believe that we've
been seeing successful results for this test because the subsequent
build in the test would fall back to a different Android SDK that was
already on disk.  The test has been improved by asserting that the newly
installed Android SDK is used during the subsequent build.