[runtime] Introduce simple local/stack string classes
This commit introduces a couple of top-level classes to work with
strings allocated either on stack or from the heap, which are always
released when leaving the scope.  The classes allocate string buffer on
stack if the size doesn't exceed a certain maximum value (provided as a
template parameter) or, if the space needed is more than size of the
on-stack buffer, allocate memory from heap as necessary.  If we choose
the stack allocation size correctly, this has the advantage of being
very quick since no calls are necessary for (de)allocation of memory -
it's either part of the function prolog/epilog where appropriate stack
for local variables space is reserved, or part of a scope setup which
operation also isn't expensive.

Two introduced classes are `static_local_string` and
`dynamic_local_string`.  The former is useful in situations when we know
the string size beforehand and we know it will never exceed that size.
The class doesn't support resizing of the buffer.  The latter adds the
ability to extend the available space, possibly transfering data from
the stack to a dynamically allocated buffer.

Both classes derive from a single base class which can be configured to
use any kind of memory storage implementation, if it's needed in the
future.

= Performance changes (app Displayed Time)

Device:
  - Model: **Pixel 3 XL**
  - Native architecture: **arm64-v8a**
  - SDK version: **29**

== Release
Release performance borders on the statistical error, but it is still
slightly better (on slower devices it will be more pronounced):

| Before  | After   | Δ        | Notes                          |
| ------- | ------- | -------- | ------------------------------ |
| 859.000 | 862.800 | +0.44% ✗ | preload enabled; 32-bit build  |
| 853.400 | 851.800 | -0.19% ✓ | preload disabled; 32-bit build |
| 867.600 | 852.200 | -1.78% ✓ | preload enabled; 64-bit build  |
| 855.000 | 858.000 | +0.35% ✗ | preload disabled; 64-bit build |

== Debug
Debug performance has bigger gains because the runtime does much more
parsing and string processing than the Release one:

| Before   | After    | Δ        | Notes                          |
| -------- | -------- | -------- | ------------------------------ |
| 1124.700 | 1113.400 | -1.00% ✓ | preload enabled; 32-bit build  |
| 1120.500 | 1114.200 | -0.56% ✓ | preload disabled; 32-bit build |
| 1122.700 | 1123.000 | +0.03% ✗ | preload enabled; 64-bit build  |
| 1113.700 | 1112.100 | -0.14% ✓ | preload disabled; 64-bit build |